retrieve Hansard lines from txt files on box.com
Retrieve the text of Hansard lines from txt files stored on box.com, using a
csv containing the filenames and line numbers to be retrieved.

This script is useful for when you've finished hand-labeling Hansard lines
and now want to use the filenames and line numbers to retrieve the full
text of each line.