move TxtParser line labeling logic to RuleLineLabeler
Moves the rule-based line labeling logic in TxtParser to RuleLineLabeler.

This provides a nicer separation of concerns, so TxtParser does not have
to worry about implementing logic for labeling a Hansard line. Instead,
it just needs a `LineLabeler` attribute, which is an instance of a class like
RuleLineLabeler that conducts the line labeling.