add overrides for `hansard_line_type4` header predictions
Add overrides for `hansard_line_type4` header predictions to address the
problem of false positive headers.

These overrides are simple and do not cover all false positives, but cover many
cases. An override is only applied if line i is a header and neither i-1 nor i+1 is a header.